Idempotence tests
